Make review files open focused previews
