+++ frontend/src/lib/utils/threadSidePanelController.test.js
@@ -4,7 +4,9 @@ import assert from 'node:assert/strict';
 import {
   buildThreadSidePanelAddMenuItems,
   createDefaultThreadSidePanelTabs,
+  filePreviewThreadSidePanelTabId,
   isThreadSidePanelSingletonKind,
+  openFilePreviewThreadSidePanelTab,
   openSingletonThreadSidePanelTab,
   THREAD_SIDE_PANEL_DEFAULT_TAB_KINDS,
   THREAD_SIDE_PANEL_SINGLETON_TAB_DEFINITIONS,
@@ -45,4 +47,29 @@ test('closed singleton tabs are restored from the side panel add menu', () => {
 
   assert.equal(next.activeTabId, 'handoff-summary');
   assert.ok(next.tabs.some((tab) => tab.kind === 'handoff-summary' && tab.label === 'Handoff'));
+});
+
+test('file preview tabs are dynamic and reused by path', () => {
+  const state = {
+    tabs: createDefaultThreadSidePanelTabs(),
+    activeTabId: 'activity',
+    nextBrowserTabIndex: 1,
+  };
+
+  const first = openFilePreviewThreadSidePanelTab(state, 'docs/diagrams/current-generation-architecture.puml', 101);
+  const second = openFilePreviewThreadSidePanelTab(first, 'docs/GENERATION_DISPATCHER_PRD.md');
+  const reused = openFilePreviewThreadSidePanelTab(second, 'docs/diagrams/current-generation-architecture.puml', 202);
+
+  assert.equal(first.activeTabId, filePreviewThreadSidePanelTabId('docs/diagrams/current-generation-architecture.puml'));
+  assert.equal(first.tabs.at(-1).kind, 'file-preview');
+  assert.equal(first.tabs.at(-1).label, 'current-generation-architecture.puml');
+  assert.equal(first.tabs.at(-1).filePath, 'docs/diagrams/current-generation-architecture.puml');
+  assert.equal(first.tabs.at(-1).runId, 101);
+  assert.equal(second.tabs.filter((tab) => tab.kind === 'file-preview').length, 2);
+  assert.equal(reused.tabs.filter((tab) => tab.kind === 'file-preview').length, 2);
+  assert.equal(reused.activeTabId, first.activeTabId);
+  assert.equal(
+    reused.tabs.find((tab) => tab.id === first.activeTabId).runId,
+    202,
+  );
 });
